requirements.txt
# ─── LLM y Embeddings (Azure OpenAI) ─────────────────────────────────────────
langchain==0.3.14
langchain-openai==0.3.1
langchain-community==0.3.14
openai==1.59.6

# ─── Vector Store ─────────────────────────────────────────────────────────────
faiss-cpu==1.9.0

# ─── Configuración y variables de entorno ─────────────────────────────────────
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1

# ─── Procesamiento de documentos ──────────────────────────────────────────────
pdfplumber==0.11.4
pdfminer.six==20231228
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
Pillow==11.1.0

# ─── Análisis y manipulación de datos ────────────────────────────────────────
pandas==2.2.3
numpy==1.26.4

# ─── Utilidades ───────────────────────────────────────────────────────────────
tqdm==4.67.1
markdown==3.7
tiktoken==0.8.0

# ─── Evaluación RAG ───────────────────────────────────────────────────────────
ragas==0.2.6
datasets==3.2.0

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0

# ─── Interfaz (opcional para demo) ───────────────────────────────────────────
streamlit==1.41.1
plotly==5.24.1
